+++ examples/readme.rs
@@ -1,15 +1,6 @@
 extern crate n5;
 
-use n5::{
-    DatasetAttributes,
-    DataBlock,
-    DataType,
-    N5Reader,
-    N5Writer,
-    VecDataBlock
-};
-use n5::compression::CompressionType;
-use n5::filesystem::N5Filesystem;
+use n5::prelude::*;
 
 fn n5_roundtrip(root_path: &str) -> std::io::Result<()> {
     let n = N5Filesystem::open_or_create(root_path)?;